Collect cluster elements and merge sub-tasks properly
This fixes two things.

First of all, the collectClusters() method used data points first "k"
data points instead of centroids as keys that identify clusters. It
also added wrong elements to these clusters, because it was using
centroid index instead of data index to get an element to add to the
cluster vector.

Doing this properly actually exposed a problem in the compute()
method, specifically in the part that merged the results of the
sub-tasks. This has been rewritten in a much more concise and
actually correct way.

This actually increases the execution time of a benchmark
repetition from approximately 800ms to 3000ms.