Fix centroid update step (compute average in KMeansBench for now)
To update centroid position, we have to calculate the average values
over all elements assigned to a particular cluster. This simplifies
average calculation and moves it to KMeansBench for now. It also gets
rid of the result vector.

The problem with computing average in the KMeansTask was that it was
only computed on merged result, but never for the directly computed
result. This would be fine, but for some vector sizes, the task would
never split and thus never compute updated centroid position.

To make it correct, the centroid update step was moved to KMeansBench
so that it is applied to the result of the "assignment to clusters"
step. The centroid update step should be also split into tasks.