Move the fork/join pool to instance level
This means that we will not be creating a new fork/join pool for
every repetition, but we will wait for quiescence after each
repetition (there should be no tasks in there anyway).

When all repetitions of the benchmark finish executing, we shut down
the pool and await its termination.