Use List (and ArrayList) instead of Vector
This drastically limits contention on Vector's elementAt() synchronized
method and improves performance by two orders of magnitude. The
contention was especially bad in this implementation, because here we
do not make copies of subvectors (we only split range of indices) for
the forked tasks. As a result, all threads read data from a single
shared Vector instance.