Rewire classloading mechanism
This commit restructures how Renaissance classloading hierarchy is
created and used.

So far, the harness was loaded via the default classloader, i.e. the
harness classes and also Scala run-time classes were directly
included in the main JAR. When the benchmark is to be loaded, new
classloader is created for each benchmark. This benchmark classloader
does not use the harness classloader as its parent to create as clean
environment as possible. Because the parent cut-off, it is not
possible to directly share core classes (e.g. the
RenaissanceBenchmark class). Thus, each benchmark also packs its own
copy of the core and harness access its via reflection.
The downside is that we have to have special proxy classes as the
classes are not directly assignable (although they are of the same
type, the have a different classloader).

The proposed solution is following. The harness is packed the same way
as a benchmark, i.e. its JAR is packed inside the main JAR. Thus, in
the main JAR, only core classes remain. Each benchmark is then loaded
in a separate classloader but it has the default classloader as its
parent. Therefore, there is no issue with assignability as the
implementation classes share common parent. To actually run the
harness, a small launcher was added to the core that creates a new
classloader for the harness.

In the end, all benchmarks and also the harness live in separate
spaces (classloaders) but they share the common interfaces from core.
This removes the need for the proxies and minimizes need for
reflection. As a side-effect, it also cleaned-up the main JAR as
Scala run-time is not included and only core classes remain. However,
these core classes must be in plain Java without any helper classes
to keep the parent (top-most) classloader clean (with only java.*
classes).

Smaller implementation details: a classloader helper is now part of
the core to avoid at least some code duplication. src/ folder is no
longer needed as it is much easier to reference core explicitly.
Otherwise the changes are minimal.

There is one issue with the modifications of the build files. Each
benchmark depends on core (as it needs the interface definitions).
However, we cannot include the core JAR in each benchmark as it would
break the assignability. So far, this is solved in a bit hackish way
inside the main build.sbt. I do not know how to fix that.

In general, this should not affect performance as the benchmarks are
still loaded in separate classloaders. There is only a small
performance penalty for the initial start-up as the harness itself is
loaded in a new classloader.