Add renaissance-core to aggregated tasks
This means that renaissance-core is included in buildFormat
and buildFormatCheck tasks, which previously ignored it.

Also by adding it to the root project, the root aggregate task
clean now properly cleans renaissance-core project.

It also appears that the renaissanceFormat and the
renaissanceFormatChecks are not unnecessary. The aggregation
seems to work as expected and even includes the renaissance build
project, which the renaissaceFormat* tasks ignored.