Add initial support for metadata in annotations
This updates the annotation interfaces in Benchmark so that what was
previously @Description is a now @Summary (one liner), and what was
previously @LongDescription is now @Description (paragraph).

It removes unnecessary methods from RenaissanceBenchmark, and
simplifies handling of benchmark metadata. The benchmark-group.txt
file is not generated anymore and the build.sbt has been also
simplified a bit by moving a chunk of metadata processing into
build.scala containing BenchmarkInfo class for build purposes.