finagle-http bugfixes (fixes #148, fixes #149)
finagle-http is now sending requests in parallel. Also, the measured
part of the benchmark now excludes starting of the threads (there is a
barrier to start all the work at once).

Note that the parameters (number of clients and number of requests) were
changed to roughly correspond to the previous performance (same order of
magnitude).

Also note that this commit changes the performance of the finagle-http
benchmark drastically.

For references, here are the issues summarized:

 * Issue 148: single requests was repeatedly waited for instead of
   sending the whole request again.
 * Issue 149: no parallel requests were actually executed.