Add Git revision information into MANIFEST.MF
The information is not (yet) used anywhere but it is useful to have it
somewhere for situations when you have the JAR (and the results) but you
are no longer sure which revision it came from.

After this change, MANIFEST.MF would contain also the following lines:

Git-Head-Commit: e895d544ee7ddbd14caf975072fd51717224b509
Git-Head-Commit-Date: 2019-07-04T16:06:00+0200
Git-Uncommitted-Changes: true

The first two are self-explanatory, the last one indicates that the
build was not done on a clean tree (i.e. uncommitted changes were
present). For releases, it must be false.