Update MarkdownGenerator to reflect changes in BenchmarkRegistyr
In addition to new interface, this also involves changing the
MarkdownGenerator so that it does not initialize the registry
at load time, but in the main() method. This in turn allows
initializing the registry using a file in the filesystem in
addition to JAR resources. In general, class initialization
should not really trigger too much code execution.