Handle the --readme option in the Launcher (for now)
MarkdownGenerator needs to be able to query suite version and
title, for which it relies on values in the RenaissanceSuite
class. However, the suite gets those values from JAR manifest,
which is not available at build time until the JAR is built.
Consequently, MarkdownGenerator cannot generate README.md
properly. Until this is resolved, launch he MarkdownGenerator
when the Launcher encounters "--readme" as the only option
on the command line.