Simplify string interpolation in MarkdownGenerator
The interpolated strings now use a map of tags that are explicitly
computed outside the interpolated strings, so we can clearly see what
tags are being used and where their value comes from.

The generator also accepts arguments to override some of the tags as
well as path to the file with benchmark metadata that can be used to
initialized the benchmark registry from filesystem (instead of JAR
resources).