Move loadBenchmark() into BenchmarkInfo and remove BenchmarkLoader
The BenchmarkInfo class has all the information needed to load the
benchmark, while the BenchmarkLoader would contain just a single
method with BenchmarkInfo parameter.

Also use the BenchmarkInfo instances instead of just names in
the harness. Once a benchmark is know to exist, no need to look it
up again all the time.