Extract ConfigParser from RenaissanceSuite
This will allow the parser to be used separately by MarkdownGenerator
to get the usage text. In addition, the ConfigParser needs to be
initialized with a map of string tags that it uses to generate the
usage text, so that MarkdownGenerator can supply values that cannot
be queried without being in a JAR file.