Update proposed method names in Benchmark
Given that we have decided that "iteration" should be used in
relation with iterative algorithm, we should get rid of the word
"iteration" from the API. The proposed alternative is (measured)
"operation", which has a setUp and tearDown phase. The operation
is executed repeatedly for a given number of repetitions.