Update JMH support, validate results and minimize generated code
Move the JmhRenaissanceBenchmark base class into "jmh" package
to avoid implicit import of the Renaissance's Benchmark class.

The generated code is kept to mininum so as to keep most of the code
(in JmhRenaissanceBenchmark) visible to an IDE to ease refactoring.

The JMH-specific BenchmarkContext is minimal at the moment and will
need to be updated once legacy methods are removed.

The measured operation keeps the benchmark result around for the
operation tear down method which validates it.