Add configurable parameters and named configurations to benchmarks
Each benchmark with configurable parameters is provided with two
other configurations in addition to the "default".

The first is "test", which is used in place of a functional test in
Travis CI and makes the benchmark workloads significantly smaller so
that one iteration can be performed quickly during CI tests.

The second is "jmh", which is used by default by the JMH wrapper.
Any parameters suitable for execution in JMH should be set there.

What is still missing:
- Listing of supported configurations and parameters from CLI.
- Writing actual parameter values to JSON output.
- Overriding parameters from CLI.