Separate configuration, plugin loading, and listener registration
Configuration will mostly deal with properly formatted strings and
let the harness create the necessary objects when it is about to run
the benchmarks.

In case of plugins, this means that we will store (reasonably well
formatted) specifications, let the harness load them and use an
event dispatcher builder to register them in listener lists.

The same now applies to ResultWriter instances. The Config just
stores output file names, the harness creates them, and uses the
event dispatcher builder to register them.

This way we ensure consistent ordering of listeners.

The loading of plugins in the ModuleLoader has been refactored a bit
to let the harness take the initiative and to deal with plugin loading
errors at a higher level.

Also, the loader only tries to construct file system URLs, because
the way we were splitting the class-path specification (on double colon)
caused URLs to be split right after the protocol.