Add support for providing arguments to plugins
The CLI now provides the `--with-arg` option, which appends an
argument to the plugin or policy specified last. The following
command line will therefore two arguments to each of the plugins:

java -jar renaissance.jar \
  --plugin <class-path>!Plugin1 --with-arg arg1.1 --with-arg arg1.2 \
  --plugin <class-path>!Plugin2 --with-arg arg2.1 --with-arg arg2.2

Plugins that want to receive command line arguments need to implement
a constructor which accepts an array of strings.