Update methods names in Benchmark to avoid the word "iteration"
This is bit of a sweeping change, hopefully we will settle down on the
names. One of the reasons was to avoid the word "iteration", which
we want to reserve for iterative algorithms and instead work with the
concept of measured operation.

The names should be familiar to JUnit 5 users.