Switch Config from Java to Scala
The moves Config to the Scala part of the harness, which deals with
harness configuration and initialization. Once the harness starts
executing the benchmarks, there is no need to access the config and
the Java part responsible for driving the execution of one benchmark
does not even know about it.