Move ExecutionPolicy into Plugin and remove registerOpetion() method
This turns execution policy into a plugin which can register for
harness events.

On the one hand, this moves responsibility for handling per-benchmark
state to the policy plugin (the harness will only create a single
instance of the policy class).

But on the other hand, it provides the policy with greater flexibility,
because it can listen to all harness events, including custom results.

It also simplifies the handling of execution policy within the harness,
because it can be (mostly) treated as yet another plugin.