Avoid unnecessary disk IO and add hash-based validation
Instead of unpacking the input file to disk letting Spark read it into
an RDD, we directly read the contents of the zipped file into a
collection of lines and build an RDD using the SparkContext's
parallelize() method.

We also avoid writing the ranks from the last executed operation
to disk.

In addition, add a hash-based validation of the results. The computed
ranks are rounded to 8 decimal places and ordered by the web site id
so as to avoid numerical instability.