Support forced GC before/after measured operation
This is initial support for forced GC. We may still need to decide
whether we want this operation to print anything and where in the
execution sequence they should be positioned.

Currently they are implemented as built-in plugins and execute after
operation set up and before operation tear down. Maybe we want the
"after" to be after operation tear down. Or we may want them outside
the repetition unix timestamps.

They also print a message between the operation started/operation
completed bracket, but maybe we want to print something outside the
bracket.

This may trigger some changes in the API (and maybe introducing
explicit before/after setup/teardown listener interfaces), depending
on what we actually need.