Add more listener types and include Before/After in type names
It appears that some plugins (e.g., forced GC) may be interested in
plugging into few other places in the execution sequence, so I have
added another set of pair events, which allows to get notified before
benchmark set up and after benchmark tear down.

This requires that the original BenchmarkSetUpListener and
BenchmarkTearDownListener listener types (which were notified after
set up and before tear down) be renamed to AfterBenchmarkSetUpListener
and BeforeBenchmarkTearDownListener, so that I could add the
BeforeBenchmarkSetUpListener and AfterBenchmarkTearDownListener types.

With this change, it is only natural to add the Before/After prefix
to other paired-event types. Consequently, the HarnessInitListener
and HarnessShutdownListener have become AfterHarnessInitListener
and BeforeHarnessShutdownListener, and the OperationSetUpListener
together with OperationTearDownListener have become
AfterOperationSetUpListener and BeforeOperationTearDown listener.

If we ever need to add BeforeOperationSetUpListener and
AfterOperationTearDownListener, at least we will not have to change
the names of the existing interfaces again.