Don't include dummy benchmarks/groups in output of CLI list commands
Removes dummy benchmarks from the output of the --list, --raw-list,
and --group-list CLI commands to remove the need to filter out these
benchmarks and groups in scripts.

This is consistent with the fact that dummy benchmarks are ignored
when the user requests to run "all" benchmarks, where "all" is
considered to be a special group containing all (public) benchmarks.