Update scala in benchmarks/actors to version 2.11.12
Keeps the dependencies untouched (yet), but adds comments to
indicate potential upgrade paths. This may actually require
splitting the akka-uct and reactors benchmarks into two sub-projects.

AkkaUct depends on akka-actor (which is still being developed).
Versions 2.5.x and 2.6.x of akka-actor support Scala up to 2.13,
version 2.3.x supports Scala only up to 2.11.

Reactors depends on reactors-core, which (even in version 0.8) only
supports Scala up to 2.11. Version 0.9 is in development, but it is
not clear which version of Scala it supports.