Update Scala to 2.12.12 and Spark to 2.4.7 in benchmarks/apache-spark
Spark version 2.4.2 and later is pre-built with Scala 2.12, unlike
the previous Spark 2 versions, which were pre-built with Scala 2.11.
This allows moving to Scala 2.12.12 to avoid a miscompilation bug in
Scala versions prior to 2.12.7.

With the update to Spark 2.4.7, we also set spark.driver.host to
"localhost" explicitly, because just setting spark.driver.bindAddress
to "127.0.0.1" was not enough to bind Spark to local host to avoid
trouble discussed in issue #66. Specifically, I was seeing failing
tasks in the page-rank benchmarks due to connections being refused.
Reverse-lookup issues may have played a role as well.