Separate benchmark group from benchmark project
The benchmark project determines a set of dependencies and we
now call it "module", to get away from the "group" concept, which
is meant to be user-visible.

Includes initial support for multiple groups.

Also includes changes to the module loader and launcher to create
scratch directory in the system temporary directory.