Rework access to benchmark parameters
This was triggered by the need to provide access to a boolean
parameter value. Instead of adding another method to BenchmarkContext,
I have modified BenchmarkContext to provide only a single method to
access benchmark parameters, which returns BenchmarkParameter instance.

The BenchmarkParameter class then provides access to the parameter
value (as string) along with a number of convenience methods that
convert the string to a typed value.

This will prevent changes to the BenchmarkContext interface whenever
we want to add a common parameter access method.

In addition, we can remove some of the (duplicated) code implementing
BenchmarkContext in the ExecutionDriver and JmhRenaissanceBenchmark.

Benchmark updates to the interface follow.