Remove the --raw-list-compatible CLI option
One less option to worry about, based on the assumption that the
--raw-list option is meant for machines (scripts), which should
get a list of compatible benchmarks by default.

Using the --no-jvm-option together with --raw-list will produce
a list of all benchmarks, regardless of compatibility.

The output of the --list option, which is meant for humans, was
modified to indicate whether a benchmark is compatible with the
current JVM.