Unify Spark parallelism parameters and query them in SparkUtil
The `thread_count` parameter is now called
`spark_executor_thread_count` to make it clear that it represents
the number of threads per Spark executor.

The number of Spark executor instances has been made into an explicit
parameter called `spark_executor_count`, which currently defaults to
4 in all benchmarks.

The setUpSparkContext() method now receives the BenchmarkContext,
which it uses to query the parameters, as well as set up Spark and
Hadoop scratch directories.