Tell SBT to use 2048 megs of memory
SBT uses 1024 by default and kills memory-related JVM options in the
`.sbtopts` file and derives its values on its own. The default memory
size results in code cache of 128 megs, which is insufficient for a
pristine build, where it can reach slightly over 200 megs. Raising the
memory to 2048 increases the code cache size to 256 megs.