Use git describe to determine Renaissance version
This will avoid having to change the version number in multiple
places. The current version number is determine by the latest
(annotated) tag and the distance to it from the current commit.
Release built from the tagged commit will have a "clean" version
number, others will include the hash of the build commit.

CI scripts will be able to determine the current version simply
by using `git describe` and stripping the initial `v`.

The only thing that needs to be updated is the git.baseVersion
setting in renaissance-core/version.sbt, which provides a base
(upcoming) version number for git-less situations, i.e., when
building the sources outside the git tree.