Use runtime dependencies and refactor generator tasks in build.sbt
Dismantles the `projectJars` and `jarsAndListGenerator` tasks which
did too many things into separate tasks responsible for generating the
`modules.properties` and `benchmarks.properties` files and for copying
the runtime dependencies into the `resource_managed` directory of the
root project. The tasks are registered separately.

The inline definition of the task to generate JMH wrappers was
factored out from the project configuration into a separate task
which also uses the `renaissanceBenchmarks` projects to select
benchmarks for which to generate the wrappers, obsoleting one
of the methods in `RenaissanceJmh`.