Build BenchmarkSuite on top of BenchmarkRegistry and ModuleLoader
The `BenchmarkSuite` class provides the harness with an interface that
allows it to (mostly) forget about the `BenchmarkRegistry` and
`ModuleLoader` classes. Both could (and should) be made package private
if it weren't for the `MarkdownWriter` class, which should really be
part of the build system, not the harness.

The `BenchmarkInfo` class has been replaced with `BenchmarkDescriptor`
which is a bit lightweight in the sense that it does not try to pull
data from `benchmark.properties` into instance variables and instead
uses the underlying map of the properties whenever we try to get some
information about a benchmark.

The `BenchmarkSuite` class provides some services that do not belong
to the `BenchmarkDescriptor` class even though, because they would
make the descriptor into more than it should be. It also collected
some of the responsibilities that were either duplicated (benchmark
context creation) or difficult to place elsewhere and were moving
between `ModuleLoader` and `BenchmarkInfo` (creating benchmark and
extension instances).

In the end, this allowed to simplify other classes such as the
`ExecutionDriver`, `RenaissanceSuite` and `JmhRenaissanceBenchmark`.

In addition, because `BenchmarkSuite` provides the `ExecutionDriver`
with an instances of `InternalBenchmarkContext`, the `ExecutionDriver`
can now print the configuration name that depends on whether there are
any overridden benchmark parameters.