Allow comparing to Optional<Version> and remove thisJvmVersion()
The `thisJvmVersion()` method is now a convenience method provided by
the `BenchmarkSuite` class -- here it was just pulling a dependency
on `ManagementFactory` into what was meant to be a simple value class.

On the other hand, the `compareTo(Optional<Version>)` comes handy and
there should be no reason to duplicate it in other classes.