Keep execution context and use SuiteBenchmarkContext in ExecutionDriver
The `ExecutionDriver` class only kept part of the execution context
in instance variables and required `RenaissanceSuite` to pass a lot
of it to `executeBenchmark()` method.

Because none of the information changes, the `ExecutionDriver` class is
now constructed with all that information and the `executeBenchmark()`
only serves to trigger the benchmark execution.

Similarly, the (internal) `executeOperation()` methods also uses most
of the captured context and is only passed the index of the operation
because that's the only thing that changes.

The `ExecutionDriver` no longer doubles as a `BenchmarkContext`. The
`SuiteBenchmarkContext` created by the `BencharkSuite` is now used in
that role, but the internal type allows querying the actual name of
the configuration which will reflect the fact that some of its
parameters were overridden.