Convert 'naive-bayes' to use Spark ML API instead of MLlib
The MLlib version of NaiveBayes in Spark 3 is a wrapper around the ML
version and the first thing the `run()` method in the MLlib version
does is convert the input RDD to DataFrame, which is quite costly to
do in each benchmark operation.

By using the newer ML API, we can simply load the DataFrame directly
without having to parse the `libsvm` format manually and use the
`fit()` method of the ML version on the DataFrame.

This makes the operation times similar to Renaissance 0.11.

Also prepares the input file the same way as `dec-tree` so that we
can factor it out later on.