Avoid conversion from RDD to DataSet in 'log-regression'
The benchmark already uses the ML API, but it loads the input file
manually as an RDD and converts the RDD to a DataFrame on every
repetition.

We now load the DataFrame directly using a Spark-provided `libsvm`
format reader and use the result directly with the `fit()` method.

We also prepare the input file the same way as `dec-tree` and
`naive-bayes` so that we can factor it out later on.