Print a message if the benchmark operation fails
This gives us a chance to at least print the name of the failure
exception, because the exception can be lost if any of the tear-down
methods in the finally blocks throws an exception.

An alternative would be to complicate the exception handling even
more to make sure the original exception does not get lost, but at
that point, the benchmark is doomed anyway.