Replace spark_thread_count with spark_thread_limit
We now configure the Spark local executor to use as many threads as
there are processors, but we limit the number to avoid scaling issues
on huge machines.

For this reason, the `spark_thread_count` parameter, which represents
a fixed number of threads to be used with the Spark local executor, is
replaced with the `spark_thread_limit` parameter, which represents the
maximum number of threads the executor will be configured with.