Report compressed OOPs mode
There is a `java.vm.compressedOopsMode` system property on newer
JVMs (certainly 11, not 8) which provides information about the
OOPs mode used by the JVM. This is a useful bit of information to
store in the benchmark result.

On JVMs which do not provide the property, the `compressed_oops_mode`
object in the `vm` part of the JSON output will be `null`.