Update scripts and MarkdownGenerator to reflect build changes
Because don't use 'sbt-assembly' anymore, we build the bundles using
the 'package' task instead of 'assembly'. Also, because we don't use
'sbt-jmh', there is no 'Jmh' configuration in the 'renaissanceJmh'
project.

To build Renaissance, just use 'package' on the root project, or
explicitly use 'renaissance/package'. To build the JMH-enabled bundle,
use 'renaissanceJmh/package'.

This also updates the JMH bundle name in the scripts.