Fix standalone mode for Dotty
In standalone mode (without modules), even if the current class loader
is a URLClassLoader (which it is on JDK8 but not on JDK11), it may not
provide enough information about the real class path. For example, on
JDK8, the class loader returns a single URL representing the jar file
passed to 'java -jar', which is no different from the value of the
'java.class.path' system property.

So we instead try to detect if the benchmark is running in standalone
mode and if yes, we build the class path using the 'Class-Path'
attribute in the jar manifest. If this fails, we try to get class path
URLs from the class loader, and in the worst case fall back to the
current class path, which should help in situations where we try to run
the benchmark completely manually, with all jar files listed on the
command line.