Avoid using partially-initialized objects in PingPong benchmarks
The benchmarks use an inner class to work around the circular
dependency between the two reactors in the PingPong-style
benchmarks, but the communication between the two reactors is
started before the inner class is fully initialized.

Because reactors execute asynchronously with respect to the main
thread, the 'pong' reactor may observe/use an unitialized (null)
reference to the 'ping' reactor, because it will start executing
before the constructor of the inner class finishes (issue #359).

This change separates the creation of the two reactors (there is
no need for one to create the other) and even though the inner
class instance reference becomes part of the closure of the event
handling lambdas, the communication is started outside the 'ping'
reactor initializer, i.e., only when inner class instance is fully
initialized.

This apparently fixes #359.