Avoid starting benchmark message exchange in reactor initializer
The message exchange in the Roundabout benchmark is started by sending
a channel reference to another reactor. We currently do this in the
receiver reactor's body (which is really an initializer executed during
reactor construction).

It is probably OK, but not necessarily obviously correct. The reason is
that we are exposing a reference to the reactor's main channel that a
human reader can reasonably expect to become first visible only after
the return of the `spawn` method.

In absence of detailed documentation on what is or is not allowed in
a reactor body, we should prefer code that is more obviously correct.
Sending the channel reference from outside the reactor body makes it
obvious that the benchmark message exchange is only started after all
reactors are done initializing.