Avoid starting communication from class constructor
Even though this has not manifested in a problem, it is generally
bad practice to expose object references from constructors. In
this particular case, initiating benchmark message exchange from
a constructor is likely to do just that.

Separating setup phase from the start of the benchmark message
exchange does not hurt anything and makes the code easier to
understand.