Exclude log4j and commons-logging from apache-spark benchmarks
This is to avoid pulling in the log4j-1.2.17.jar. The functionality
should be replaced by reload4j, which gets pulled in by slf4j-log4j12
starting with version 1.7.36. Because reload4j did not find Spark's
default logging configuration file (log4j.properties), it was placed in
the apache-spark benchmark resource directory so that reload4j can find
it. This appears to provide a reasonable logging configuration again.

The commons-logging package should be subsumed by the jcl-over-slf4j
logging bridge.