CI: use versioned image names
So far, images from the docker.io/renaissancebench/buildenv were
versioned by JDK version only. That makes testing of new images
impractical as there is no way to distinguish releases of the
images (e.g. the image name does not reflect which Fedora version
was used).

The v1- version is a tagged version of existing images and hence
the transition should be smooth. v2- would come in a separate
commit and would probably introduce images built on new Fedora.