Temporarily restrict supported JVM version in several benchmarks
Changes in JDK21 break apache-spark benchmarks and the scala-dotty
benchmark (see #370 and #384). With the apache-spark benchmarks, we
are waiting for a fix that is expected to appear in Spark 3.5. With
the scala-dotty benchmark, we need to check if the issue is handled
upstream. For the time being (to allow including JDK21 in the CI
testing matrix), we will restrict the supported JVM in these benchmark
to JDK20.