Migrate scala-dotty to Scala 3
This should avoid issues with using Scala 3 code (scala3-compiler_3)
from Scala 2. This broke with version 3.1.3 of scala3-compiler when
Scala 2.13 compiler start to complain about "Unsupported Scala 3 union
in parameter value" (3.1.2 was the last to work).

For compatibility with JDK21, we need fixes to Tasty reader which was
broken by changes to javac in JDK21-b21 (see JDK-8292275), which only
made it into version 3.3.1-RC1.

Using reflection would probably fix the problem, as might a Java-based
bridge method, but switching to Scala 3 seems to work as well and
paves the way to introducing more Scala 3 benchmarks in the future.