Extract JSON loading to Neo4jAnalytics and use play-json
This moves the dependency on JSON library from the AnalyticsBenchmark
class to the Neo4jAnalytics class, making the AnalyticsBenchmark
depend only case class model of the graph data.

Also switch from lift-json to play-json, because Scala 2.12.18
requires newer scala-xml which conflicts with the version required
by lift-json. The play-json library also appears to be more alive.