Preserve order of dependencies when generating MANIFEST.MF
The order of dependencies stored in MANIFEST.MF can be different
from what is stored in benchmark.properties. This generally does not
hurt, except in the case of the `scala-dotty` benchmark.

In this particular case, we need to include a Scala 2.13 version of
`scala-compiler` to compile the `scalap` (workload) code. However, if
the Scala 2.13 jar ends up before `scala3-compiler_3` on the class path
stored in MANIFEST.MF, the compiler will use some classes from the
Scala 2.13 jar and the produced Tasty files will have a different
header. Because the reordering could only happen when generating the
classpath for MANIFEST.MF, running the benchmark in standalone mode
resulted in a failure, even though the benchmark ran fine when run
from the bundle.