Use invoke() on task to avoid potential NullPointerException
The getPool() method of may return null if the current thread is not
hosted by any ForkJoinPool. This can apparently happen when calling
fork() from a thread is not a ForkJoinWorkerThread executing as a
ForkJoinPool computation.

This can apparently happen when executing under JMH with the -t option,
but the real reason may be calling pool.invoke(task) from a thread that
is not a worker itself.