Set/restore context class loader around benchmark methods in JMH
Wraps `Benchmark` interface methods with code that set and restores the
context class loader for the current thread when running with JMH. This
is because JMH can be asked to create more benchmark runners that
invoke the `Benchmark.run()` method concurrently but do not replicate
the context class loader of the (initial) main thread.

Even though this is not really supported (or intended) use of the
`Benchmark` interface, the benchmarks should not crash because they
fail to load their classes.