Make the number of "camera scans" constant for a given workload
Camera scans are requested based on thread progress. To keep the number
of camera scans constant, the requests are queued to avoid potential
loss of wake-ups when the rate of progress increases (due to reduced
contention towards the end of the workload).

This is a proof-of-concept, it would be nice to replace the
LinkedBlockingQueue with an STM-based solution.