Use Spark to load CSV files and avoid extra IO
Use the Spark DataFrameReader to load the CSV files with movies and
ratings to correctly load CSV files with quoted values (movies). There
is also no need to copy the resource files to the scratch directory to
avoid extra IO.

Given the above changes, the data initialization was made more localized
to avoid keeping data that are not needed for benchmark execution after
the initial setup.

When loading the data using the DataFrameReader from Spark Session
instead of the textfile() from Spark Context (which uses hadoopFile()
underneath), the data is partitioned differently, which seems to affect
performance. To mimic the original behavior, we repartition the data
with the same defaults used by hadoopFile().