Trigger early loading of the UserGroupInformation.RealUser class to avoid crash on exit
One of Hadoop shutdown hooks apparently accesses the class, but at that
time, the class loader providing access to the benchmark's dependency
JARs is already closed and the files deleted, resulting in a class
loading exception in the shutdown handler. Loading the class early
ensures that the class will be available to the late shutdown hook.