Patch Spark Unsafe jar to avoid looking up non-existing internal class since Java 26-ea+6
Reuses the infrastructure used for patching the Hadoop Client API and
moves the patching and remapping from the main 'renaissance' project to
the 'apacheSparkBenchmarks' project, which removes the need for doing
this twice in the 'renaissance' and 'renaissanceJmh' projects.

Also moves most of the patching logic from build.sbt into patcher.scala
(previous hadoop.scala) to reduce clutter.