Do not close the harness class loaders to avoid exceptions in shutdown hooks
Some of the frameworks register shutdown hooks that may load classes
that were not loaded during benchmark run. Closing the class loaders
results in ClassNotFound exceptions behing thrown in their shutdown
hooks, which is rather annoying. However, there is no way to ensure that
our shutdown hook runs as the very last thing, so it is better to keep
the class loaders open.